+++ lib/define_algo_order.js
@@ -2,7 +2,7 @@
 
 const _isObject = require('lodash/isObject')
 const defaultOrderErrorHandler = require('./default_handlers/orders_order_error')
-const defaultErrorInsufficientBalanceHandler = require('./default_handlers/error_insufficient_Balance')
+const defaultErrorInsufficientBalanceHandler = require('./default_handlers/error_insufficient_balance')
 const defaultErrorMinimumSizeHandler = require('./default_handlers/error_minimum_size')
 
 /**
